Optimize `WordPress/Resources/` via ImageOptim
I used ImageOptim version 1.8.8. The result was:

> Saved 1.9 MB out of 8.8 MB. 22.2% overall (up to 91.4% per file)

This is a followup to the work in
8d09f32a22f21d58b0f9cb7f135d7f25f782799d, which remove the alpha channel
from the App Store icon only. That commit was sharp focused because it
was made onto the release branch.

After these changes, the following command returns no matches:

```
sips -g hasAlpha WordPress/Resources/Icons/**/*.png | grep -B1 'yes'
```